Refactoring signals and termination

+++ src/main/java/com/spotify/confidence/eventsender/EventSenderEngineImpl.java
@@ -9,7 +9,7 @@ class EventSenderEngineImpl implements EventSenderEngine {
   private final ExecutorService writeThread = Executors.newSingleThreadExecutor();
   private final ExecutorService uploadThread = Executors.newSingleThreadExecutor();
   private final BlockingQueue<Event> writeQueue = new LinkedBlockingQueue<>();
-  private final BlockingQueue<String> uploadQueue = new LinkedBlockingQueue<>();
+  private final BlockingQueue<Object> uploadQueue = new LinkedBlockingQueue<>();
   private final EventSenderStorage eventStorage = new InMemoryStorage();
   private final List<FlushPolicy> flushPolicies;
   private boolean isStopped = false;
@@ -32,8 +32,7 @@ class EventSenderEngineImpl implements EventSenderEngine {
           if (flushPolicies.stream().anyMatch(FlushPolicy::shouldFlush)) {
             flushPolicies.forEach(FlushPolicy::reset);
             eventStorage.batch();
-            System.out.println("UPLOAD SIGNAL");
-            uploadQueue.add("Upload");
+            uploadQueue.add(new Object());
           }
         } catch (InterruptedException e) {
           throw new RuntimeException(e);
@@ -46,7 +45,7 @@ class EventSenderEngineImpl implements EventSenderEngine {
     private final EventUploader uploader =
         event -> {
           System.out.println(
-              "Event uploader -> batch -> "
+              "Sending batch: "
                   + event.events().stream().map(e -> e.name).collect(Collectors.toList()));
           return CompletableFuture.completedFuture(true);
         };
@@ -55,18 +54,17 @@ class EventSenderEngineImpl implements EventSenderEngine {
     public void run() {
       while (!isStopped) {
         try {
-          System.out.println("UPLOAD SIGNAL Waiting");
           uploadQueue.take();
-          System.out.println("UPLOAD SIGNAL COMING");
           List<EventBatch> batches = List.copyOf(eventStorage.readyEvents());
-          // for each batch upload -> boolean should clean up ->
+          if (batches.isEmpty()) {
+            continue;
+          }
           for (EventBatch batch : batches) {
             boolean uploadSuccessful = uploader.upload(batch).get();
             if (uploadSuccessful) {
               eventStorage.deleteBatch(batch.id());
             }
           }
-          System.out.println("Finished upload");
         } catch (InterruptedException | ExecutionException e) {
           throw new RuntimeException(e);
         }
@@ -86,14 +84,18 @@ class EventSenderEngineImpl implements EventSenderEngine {
 
   @Override
   public void close() {
-    eventStorage.batch();
-    uploadQueue.add("UPLOAD");
-    try {
-      writeThread.awaitTermination(0, TimeUnit.SECONDS);
-      uploadThread.awaitTermination(6, TimeUnit.SECONDS);
-    } catch (InterruptedException e) {
-      e.printStackTrace();
-    }
+    // Prevent runnables from waiting on the blocking queues
     isStopped = true;
+    // Trigger the runnables one more time if waiting on blocking queues
+    uploadQueue.add(new Object());
+    writeThread.shutdown();
+    uploadThread.shutdown();
+    try {
+      boolean isUploadTerminated = uploadThread.awaitTermination(10, TimeUnit.SECONDS);
+      System.out.printf(
+          "Termination complete, upload thread correct termination %S%n", isUploadTerminated);
+    } catch (InterruptedException e) {
+      System.out.println(e.getMessage());
+    }
   }
 }
